fix: refactor config creator into command and implementations
