+++ src/implementations/config/__tests__/expected-config-file-using-insert-file-option.json
