+++ src/implementations/config/__tests__/expected-config-file-using-type-option.json
